requirements.txt
streamlit
pandas
seaborn
plotly.express
matplotlib.pyplot
plotly.graph_objects
statsmodels.tsa.seasonal
hugchat==0.0.8